split batch2 into batch2 (front end) and batch2_driver
