+++ batch2/Dockerfile
@@ -8,5 +8,3 @@ RUN pip3 install --no-cache-dir /batch && \
   rm -rf /batch
 
 EXPOSE 5000
-
-CMD ["python3", "-m", "batch"]
