+++ batch2/deployment.yaml
@@ -6,19 +6,19 @@ metadata:
 apiVersion: apps/v1beta2
 kind: Deployment
 metadata:
-  name: batch2
+  name: batch2_driver
   labels:
-    app: batch2
+    app: batch2_driver
     hail.is/sha: "{{ code.sha }}"
 spec:
   selector:
     matchLabels:
-      app: batch2
+      app: batch2_driver
   replicas: 1
   template:
     metadata:
       labels:
-        app: batch2
+        app: batch2_driver
         hail.is/sha: "{{ code.sha }}"
         grafanak8sapp: "true"
     spec:
@@ -27,8 +27,12 @@ spec:
       priorityClassName: production
 {% endif %}
       containers:
-      - name: batch2
+      - name: batch2_driver
         image: {{ batch2_image.image }}
+	command:
+	 - python3
+	 - -m
+	 - batch.driver
         resources:
           requests:
             memory: "1G"
@@ -68,7 +72,100 @@ spec:
            mountPath: /session-secret-key
            readOnly: true
          - name: batch2-user-secret
-           mountPath: /batch-user-secret
+           mountPath: /sql-config
+           readOnly: true
+         - name: gsa-key
+           mountPath: /sql-config
+           readOnly: true
+         - name: batch2-tokens
+           mountPath: /user-tokens
+           readOnly: true
+        readinessProbe:
+          httpGet:
+            path: /healthcheck
+            port: 5000
+          initialDelaySeconds: 5
+          periodSeconds: 5
+      volumes:
+       - name: deploy-config
+         secret:
+           optional: false
+           secretName: deploy-config
+       - name: session-secret-key
+         secret:
+           optional: false
+           secretName: session-secret-key
+       - name: sql-config
+         secret:
+           optional: false
+           secretName: "{{ batch2_database.user_secret_name }}"
+       - name: gsa-key
+         secret:
+           optional: false
+           secretName: batch2-gsa-key
+       - name: batch2-tokens
+         secret:
+           optional: false
+           secretName: batch2-tokens
+      tolerations:
+      - key: "preemptible"
+        operator: "Equal"
+        value: "true"
+        effect: "NoSchedule"
+---
+apiVersion: apps/v1beta2
+kind: Deployment
+metadata:
+  name: batch2
+  labels:
+    app: batch2
+    hail.is/sha: "{{ code.sha }}"
+spec:
+  selector:
+    matchLabels:
+      app: batch2
+  replicas: 1
+  template:
+    metadata:
+      labels:
+        app: batch2
+        hail.is/sha: "{{ code.sha }}"
+        grafanak8sapp: "true"
+    spec:
+{% if deploy %}
+      priorityClassName: production
+{% endif %}
+      containers:
+      - name: batch2
+        image: {{ batch2_image.image }}
+	command:
+	 - python3
+	 - -m
+	 - batch.front_end
+        resources:
+          requests:
+            memory: "1G"
+            cpu: "1"
+        env:
+         - name: HAIL_DOMAIN
+           value: "{{ global.domain }}"
+         - name: HAIL_DEPLOY_CONFIG_FILE
+           value: /deploy-config/deploy-config.json
+{% if deploy %}
+         - name: HAIL_INSTANCE_ID
+           value: cd50b95a89914efb897965a5e982a29d
+{% endif %}
+        ports:
+         - containerPort: 5000
+        volumeMounts:
+         - name: deploy-config
+           mountPath: /deploy-config
+           readOnly: true
+         - name: session-secret-key
+           mountPath: /session-secret-key
+           readOnly: true
+         - name: sql-config
+           mountPath: /sql-config
            readOnly: true
          - name: gsa-key
            mountPath: /batch-gsa-key
@@ -91,7 +188,7 @@ spec:
          secret:
            optional: false
            secretName: session-secret-key
-       - name: batch2-user-secret
+       - name: sql-config
          secret:
            optional: false
            secretName: "{{ batch2_database.user_secret_name }}"
@@ -122,3 +219,17 @@ spec:
     targetPort: 5000
   selector:
     app: batch2
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: batch2_driver
+  labels:
+    app: batch2_driver
+spec:
+  ports:
+  - port: 80
+    protocol: TCP
+    targetPort: 5000
+  selector:
+    app: batch2_driver
